Split Database into multiple files
