+++ FeistyDB/Database.swift
@@ -614,420 +614,6 @@ extension Database {
 }
 
 extension Database {
-	/// A comparator for `String` objects.
-	///
-	/// - parameter lhs: The left-hand operand
-	/// - parameter rhs: The right-hand operand
-	///
-	/// - returns: The result of comparing `lhs` to `rhs`
-	public typealias StringComparator = (_ lhs: String, _ rhs: String) -> ComparisonResult
-
-	/// Adds a custom collation function.
-	///
-	/// ```swift
-	/// try db.addCollation("localizedCompare", { (lhs, rhs) -> ComparisonResult in
-	///     return lhs.localizedCompare(rhs)
-	/// })
-	/// ```
-	///
-	/// - parameter name: The name of the custom collation sequence
-	/// - parameter block: A string comparison function
-	///
-	/// - throws: An error if the collation function couldn't be added
-	public func addCollation(_ name: String, _ block: @escaping StringComparator) throws {
-		let function_ptr = UnsafeMutablePointer<StringComparator>.allocate(capacity: 1)
-		function_ptr.initialize(to: block)
-		guard sqlite3_create_collation_v2(db, name, SQLITE_UTF8, function_ptr, { (context, lhs_len, lhs_data, rhs_len, rhs_data) -> Int32 in
-			// Have total faith that SQLite will pass valid parameters and use unsafelyUnwrapped
-			let lhs = String(bytesNoCopy: UnsafeMutableRawPointer(mutating: lhs_data.unsafelyUnwrapped), length: Int(lhs_len), encoding: .utf8, freeWhenDone: false).unsafelyUnwrapped
-			let rhs = String(bytesNoCopy: UnsafeMutableRawPointer(mutating: rhs_data.unsafelyUnwrapped), length: Int(rhs_len), encoding: .utf8, freeWhenDone: false).unsafelyUnwrapped
-
-			// Cast context to the appropriate type and call the comparator
-			let function_ptr = context.unsafelyUnwrapped.assumingMemoryBound(to: StringComparator.self)
-			let result = function_ptr.pointee(lhs, rhs)
-			return Int32(result.rawValue)
-		}, { context in
-			let function_ptr = context.unsafelyUnwrapped.assumingMemoryBound(to: StringComparator.self)
-			function_ptr.deinitialize(count: 1)
-			function_ptr.deallocate()
-		}) == SQLITE_OK else {
-			throw SQLiteError("Error adding collation sequence \"\(name)\"", takingDescriptionFromDatabase: db)
-		}
-	}
-
-	/// Removes a custom collation function.
-	///
-	/// - parameter name: The name of the custom collation sequence
-	///
-	/// - throws: An error if the collation function couldn't be removed
-	public func removeCollation(_ name: String) throws {
-		guard sqlite3_create_collation_v2(db, name, SQLITE_UTF8, nil, nil, nil) == SQLITE_OK else {
-			throw SQLiteError("Error removing collation sequence \"\(name)\"", takingDescriptionFromDatabase: db)
-		}
-	}
-}
-
-/// Passes `value` to the appropriate `sqlite3_result` function
-///
-/// - parameter sqlite_context: An `sqlite3_context *` object
-/// - parameter value: The value to pass
-func set_sqlite3_result(_ sqlite_context: OpaquePointer!, value: DatabaseValue) {
-	switch value {
-	case .integer(let i):
-		sqlite3_result_int64(sqlite_context, i)
-	case .float(let f):
-		sqlite3_result_double(sqlite_context, f)
-	case .text(let t):
-		sqlite3_result_text(sqlite_context, t, -1, SQLITE_TRANSIENT)
-	case .blob(let b):
-		b.withUnsafeBytes { bytes in
-			sqlite3_result_blob(sqlite_context, bytes.baseAddress, Int32(b.count), SQLITE_TRANSIENT)
-		}
-	case .null:
-		sqlite3_result_null(sqlite_context)
-	}
-}
-
-extension Database {
-	/// A custom SQL function.
-	///
-	/// - parameter values: The SQL function parameters
-	///
-	/// - throws: `Error`
-	///
-	/// - returns: The result of applying the function to `values`
-	public typealias SQLFunction = (_ values: [DatabaseValue]) throws -> DatabaseValue
-
-	/// Custom SQL function flags
-	///
-	/// - seealso: [Function Flags](https://www.sqlite.org/c3ref/c_deterministic.html)
-	public struct SQLFunctionFlags: OptionSet {
-		public let rawValue: Int
-
-		public init(rawValue: Int) {
-			self.rawValue = rawValue
-		}
-
-		/// The function gives the same output when the input parameters are the same
-		public static let deterministic = SQLFunctionFlags(rawValue: 1 << 0)
-		/// The function may only be invoked from top-level SQL, and cannot be used in views or triggers
-		/// nor in schema structures such as `CHECK` constraints, `DEFAULT` clauses, expression indexes, partial indexes, or generated columns
-		public static let directOnly = SQLFunctionFlags(rawValue: 1 << 1)
-		/// Indicates to SQLite that a function may call `sqlite3_value_subtype() `to inspect the sub-types of its arguments
-		public static let subtype = SQLFunctionFlags(rawValue: 1 << 2)
-		/// The function is unlikely to cause problems even if misused.
-		/// An innocuous function should have no side effects and should not depend on any values other than its input parameters.
-		public static let innocuous = SQLFunctionFlags(rawValue: 1 << 3)
-	}
-}
-
-extension Database.SQLFunctionFlags {
-	/// Returns the value of `self` using SQLite's flag values
-	func asSQLiteFlags() -> Int32 {
-		var flags: Int32 = 0
-
-		if contains(.deterministic) {
-			flags |= SQLITE_DETERMINISTIC
-		}
-		if contains(.directOnly) {
-			flags |= SQLITE_DIRECTONLY
-		}
-		if contains(.subtype) {
-			flags |= SQLITE_SUBTYPE
-		}
-		if contains(.innocuous) {
-			flags |= SQLITE_INNOCUOUS
-		}
-
-		return flags
-	}
-}
-
-/// A custom SQL aggregate function.
-public protocol SQLAggregateFunction {
-	/// Invokes the aggregate function for one or more values in a row.
-	///
-	/// - parameter values: The SQL function parameters
-	///
-	/// - throws: `Error`
-	func step(_ values: [DatabaseValue]) throws
-
-	/// Returns the current value of the aggregate function.
-	///
-	/// - note: This should also reset any function context to defaults.
-	///
-	/// - throws: `Error`
-	///
-	/// - returns: The current value of the aggregate function.
-	func final() throws -> DatabaseValue
-}
-
-/// A custom SQL aggregate window function.
-public protocol SQLAggregateWindowFunction: SQLAggregateFunction {
-	/// Invokes the inverse aggregate function for one or more values in a row.
-	///
-	/// - parameter values: The SQL function parameters
-	///
-	/// - throws: `Error`
-	func inverse(_ values: [DatabaseValue]) throws
-
-	/// Returns the current value of the aggregate window function.
-	///
-	/// - throws: `Error`
-	///
-	/// - returns: The current value of the aggregate window function.
-	func value() throws -> DatabaseValue
-}
-
-extension Database {
-	/// Adds a custom SQL scalar function.
-	///
-	/// For example, a localized uppercase scalar function could be implemented as:
-	/// ```swift
-	/// try db.addFunction("localizedUppercase", arity: 1) { values in
-	///     let value = values.first.unsafelyUnwrapped
-	///     switch value {
-	///     case .text(let s):
-	///         return .text(s.localizedUppercase())
-	///     default:
-	///         return value
-	///     }
-	/// }
-	/// ```
-	///
-	/// - parameter name: The name of the function
-	/// - parameter arity: The number of arguments the function accepts
-	/// - parameter flags: Flags affecting the function's use by SQLite
-	/// - parameter block: A closure that returns the result of applying the function to the supplied arguments
-	///
-	/// - throws: An error if the SQL scalar function couldn't be added
-	///
-	/// - seealso: [Create Or Redefine SQL Functions](https://sqlite.org/c3ref/create_function.html)
-	public func addFunction(_ name: String, arity: Int = -1, flags: SQLFunctionFlags = [.deterministic, .directOnly], _ block: @escaping SQLFunction) throws {
-		let function_ptr = UnsafeMutablePointer<SQLFunction>.allocate(capacity: 1)
-		function_ptr.initialize(to: block)
-
-		let function_flags = SQLITE_UTF8 | flags.asSQLiteFlags()
-		guard sqlite3_create_function_v2(db, name, Int32(arity), function_flags, function_ptr, { sqlite_context, argc, argv in
-			let context = sqlite3_user_data(sqlite_context)
-			let function_ptr = context.unsafelyUnwrapped.assumingMemoryBound(to: SQLFunction.self)
-
-			let args = UnsafeBufferPointer(start: argv, count: Int(argc))
-			let arguments = args.map { DatabaseValue($0.unsafelyUnwrapped) }
-
-			do {
-				set_sqlite3_result(sqlite_context, value: try function_ptr.pointee(arguments))
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, nil, nil, { context in
-			let function_ptr = context.unsafelyUnwrapped.assumingMemoryBound(to: SQLFunction.self)
-			function_ptr.deinitialize(count: 1)
-			function_ptr.deallocate()
-		}) == SQLITE_OK else {
-			throw SQLiteError("Error adding SQL scalar function \"\(name)\"", takingDescriptionFromDatabase: db)
-		}
-	}
-
-	/// Adds a custom SQL aggregate function.
-	///
-	/// For example, an integer sum aggregate function could be implemented as:
-	/// ```swift
-	/// class IntegerSumAggregateFunction: SQLAggregateFunction {
-	///     func step(_ values: [DatabaseValue]) throws {
-	///         let value = values.first.unsafelyUnwrapped
-	///         switch value {
-	///             case .integer(let i):
-	///                 sum += i
-	///             default:
-	///                 throw DatabaseError("Only integer values supported")
-	///         }
-	///     }
-	///
-	///     func final() throws -> DatabaseValue {
-	///         defer {
-	///             sum = 0
-	///         }
-	///         return DatabaseValue(sum)
-	///     }
-	///
-	///     var sum: Int64 = 0
-	/// }
-	/// ```
-	///
-	/// - parameter name: The name of the aggregate function
-	/// - parameter arity: The number of arguments the function accepts
-	/// - parameter flags: Flags affecting the function's use by SQLite
-	/// - parameter aggregateFunction: An object defining the aggregate function
-	///
-	/// - throws:  An error if the SQL aggregate function can't be added
-	///
-	/// - seealso: [Create Or Redefine SQL Functions](https://sqlite.org/c3ref/create_function.html)
-	public func addAggregateFunction(_ name: String, arity: Int = -1, flags: SQLFunctionFlags = [.deterministic, .directOnly], _ function: SQLAggregateFunction) throws {
-		// function must live until the xDelete function is invoked; store it as a +1 object in context
-		let context = Unmanaged.passRetained(function as AnyObject).toOpaque()
-
-		let function_flags = SQLITE_UTF8 | flags.asSQLiteFlags()
-		guard sqlite3_create_function_v2(db, name, Int32(arity), function_flags, context, nil, { sqlite_context, argc, argv in
-			let context = sqlite3_user_data(sqlite_context)
-			let function = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).takeUnretainedValue() as! SQLAggregateFunction
-
-			let args = UnsafeBufferPointer(start: argv, count: Int(argc))
-			let arguments = args.map { DatabaseValue($0.unsafelyUnwrapped) }
-
-			do {
-				try function.step(arguments)
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, { sqlite_context in
-			let context = sqlite3_user_data(sqlite_context)
-			let function = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).takeUnretainedValue() as! SQLAggregateFunction
-
-			do {
-				set_sqlite3_result(sqlite_context, value: try function.final())
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, { context in
-			// Balance the +1 retain above
-			Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).release()
-		}) == SQLITE_OK else {
-			throw SQLiteError("Error adding SQL aggregate function \"\(name)\"", takingDescriptionFromDatabase: db)
-		}
-	}
-
-	/// Adds a custom SQL aggregate window function.
-	///
-	/// For example, an integer sum aggregate window function could be implemented as:
-	/// ```swift
-	/// class IntegerSumAggregateWindowFunction: SQLAggregateWindowFunction {
-	///     func step(_ values: [DatabaseValue]) throws {
-	///         let value = values.first.unsafelyUnwrapped
-	///         switch value {
-	///             case .integer(let i):
-	///                 sum += i
-	///             default:
-	///                 throw DatabaseError("Only integer values supported")
-	///         }
-	///     }
-	///
-	///     func inverse(_ values: [DatabaseValue]) throws {
-	///         let value = values.first.unsafelyUnwrapped
-	///         switch value {
-	///             case .integer(let i):
-	///                 sum -= i
-	///             default:
-	///                 throw DatabaseError("Only integer values supported")
-	///         }
-	///     }
-	///
-	///     func value() throws -> DatabaseValue {
-	///         return DatabaseValue(sum)
-	///     }
-	///
-	///     func final() throws -> DatabaseValue {
-	///         defer {
-	///             sum = 0
-	///         }
-	///         return DatabaseValue(sum)
-	///     }
-	///
-	///     var sum: Int64 = 0
-	/// }
-	/// ```
-	///
-	/// - parameter name: The name of the aggregate window function
-	/// - parameter arity: The number of arguments the function accepts
-	/// - parameter flags: Flags affecting the function's use by SQLite
-	/// - parameter aggregateWindowFunction: An object defining the aggregate window function
-	///
-	/// - throws:  An error if the SQL aggregate window function can't be added
-	///
-	/// - seealso: [User-Defined Aggregate Window Functions](https://sqlite.org/windowfunctions.html#udfwinfunc)
-	public func addAggregateWindowFunction(_ name: String, arity: Int = -1, flags: SQLFunctionFlags = [.deterministic, .directOnly], _ function: SQLAggregateWindowFunction) throws {
-		// function must live until the xDelete function is invoked; store it as a +1 object in context
-		let context = Unmanaged.passRetained(function as AnyObject).toOpaque()
-
-		let function_flags = SQLITE_UTF8 | flags.asSQLiteFlags()
-		guard sqlite3_create_window_function(db, name, Int32(arity), function_flags, context, { sqlite_context, argc, argv in
-			let context = sqlite3_user_data(sqlite_context)
-			let function = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).takeUnretainedValue() as! SQLAggregateWindowFunction
-
-			let args = UnsafeBufferPointer(start: argv, count: Int(argc))
-			let arguments = args.map { DatabaseValue($0.unsafelyUnwrapped) }
-
-			do {
-				try function.step(arguments)
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, { sqlite_context in
-			let context = sqlite3_user_data(sqlite_context)
-			let function = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).takeUnretainedValue() as! SQLAggregateWindowFunction
-
-			do {
-				set_sqlite3_result(sqlite_context, value: try function.final())
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, { sqlite_context in
-			let context = sqlite3_user_data(sqlite_context)
-			let function = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).takeUnretainedValue() as! SQLAggregateWindowFunction
-
-			do {
-				set_sqlite3_result(sqlite_context, value: try function.value())
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, { sqlite_context, argc, argv in
-			let context = sqlite3_user_data(sqlite_context)
-			let function = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).takeUnretainedValue() as! SQLAggregateWindowFunction
-
-			let args = UnsafeBufferPointer(start: argv, count: Int(argc))
-			let arguments = args.map { DatabaseValue($0.unsafelyUnwrapped) }
-
-			do {
-				try function.inverse(arguments)
-			}
-
-			catch let error {
-				sqlite3_result_error(sqlite_context, "\(error)", -1)
-			}
-		}, { context in
-			// Balance the +1 retain above
-			Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(context.unsafelyUnwrapped)).release()
-		}) == SQLITE_OK else {
-			throw SQLiteError("Error adding SQL aggregate window function \"\(name)\"", takingDescriptionFromDatabase: db)
-		}
-	}
-
-	/// Removes a custom SQL scalar, aggregate, or window function.
-	///
-	/// - parameter name: The name of the custom SQL function
-	/// - parameter arity: The number of arguments the custom SQL function accepts
-	///
-	/// - throws: An error if the SQL function couldn't be removed
-	public func removeFunction(_ name: String, arity: Int = -1) throws {
-		guard sqlite3_create_function_v2(db, name, Int32(arity), SQLITE_UTF8, nil, nil, nil, nil, nil) == SQLITE_OK else {
-			throw SQLiteError("Error removing SQL function \"\(name)\"", takingDescriptionFromDatabase: db)
-		}
-	}
-}
-
-extension Database {
 	/// A hook called when a database transaction is committed.
 	///
 	/// - returns: `true` if the commit operation is allowed to proceed, `false` otherwise
@@ -1336,264 +922,5 @@ extension Database {
 		}
 
 		return (Int(current), Int(highwater))
-	}
-}
-
-/// Returns a pointer to the `fts5_api` structure for `db`.
-///
-/// - parameter db: The database connection to query
-///
-/// - throws:  An error if the `fts5_api` structure couldn't be retrieved
-///
-/// - returns: A pointer to the global `fts5_api` structure for `db`
-func get_fts5_api(for db: SQLiteDatabaseConnection) throws -> UnsafePointer<fts5_api> {
-	var stmt: SQLitePreparedStatement? = nil
-	let sql = "SELECT fts5(?1);"
-	guard sqlite3_prepare_v2(db, sql, -1, &stmt, nil) == SQLITE_OK else {
-		throw SQLiteError("Error preparing SQL \"\(sql)\"", takingDescriptionFromDatabase: db)
-	}
-
-	defer {
-		sqlite3_finalize(stmt)
-	}
-
-	var api_ptr: UnsafePointer<fts5_api>?
-	guard sqlite3_bind_pointer(stmt, 1, &api_ptr, "fts5_api_ptr", nil) == SQLITE_OK else {
-		throw SQLiteError("Error binding FTS5 API pointer", takingDescriptionFromStatement: stmt!)
-	}
-
-	guard sqlite3_step(stmt) == SQLITE_ROW else {
-		throw SQLiteError("Error retrieving FTS5 API pointer", takingDescriptionFromStatement: stmt!)
-	}
-
-	guard let api = api_ptr else {
-		throw DatabaseError("FTS5 not available")
-	}
-
-	return api
-}
-
-// Protocol declarations can't be nested, otherwise this would be inside Database
-
-/// An interface to a custom FTS5 tokenizer.
-public protocol FTS5Tokenizer {
-	/// Initializes an FTS5 tokenizer.
-	///
-	/// - parameter arguments: The tokenizer arguments used to create the FTS5 table.
-	init(arguments: [String])
-
-	/// Sets the text to be tokenized.
-	///
-	/// - parameter text: The text to be tokenized.
-	/// - parameter reason: The reason tokenization is being requested.
-	func setText(_ text: String, reason: Database.FTS5TokenizationReason)
-
-	/// Advances the tokenizer to the next token.
-	///
-	/// - returns: `true` if a token was found, `false` otherwise
-	func advance() -> Bool
-
-	/// Returns the current token.
-	///
-	/// - returns: The current token or `nil` if none
-	func currentToken() -> String?
-
-	/// Copies the current token in UTF-8 to the supplied buffer.
-	///
-	/// - parameter buffer: A buffer to receive the current token encoded in UTF-8
-	/// - parameter capacity: The number of bytes availabe in `buffer`
-	///
-	/// - throws: An error if `buffer` has insufficient capacity for the token
-	///
-	/// - returns: The number of bytes written to `buffer`
-	func copyCurrentToken(to buffer: UnsafeMutablePointer<UInt8>, capacity: Int) throws -> Int
-}
-
-extension Database {
-	/// Glue for creating a generic Swift type in a C callback
-	final class FTS5TokenizerCreator {
-		/// The constructor closure
-		let construct: (_ arguments : [String]) throws -> FTS5Tokenizer
-
-		/// Creates a new FTS5TokenizerCreator.
-		///
-		/// - parameter construct: A closure that creates the tokenizer
-		init(_ construct: @escaping (_ arguments: [String]) -> FTS5Tokenizer)
-		{
-			self.construct = construct
-		}
-	}
-
-	/// The reasons FTS5 will request tokenization
-	public enum FTS5TokenizationReason {
-		/// A document is being inserted into or removed from the FTS table
-		case document
-		/// A `MATCH` query is being executed against the FTS index
-		case query
-		/// Same as `query`, except that the bareword or quoted string is followed by a `*` character
-		case prefix
-		/// The tokenizer is being invoked to satisfy an `fts5_api.xTokenize()` request made by an auxiliary function
-		case aux
-	}
-
-	/// Adds a custom FTS5 tokenizer.
-	///
-	/// For example, a word tokenizer using CFStringTokenizer could be implemented as:
-	/// ```swift
-	/// class WordTokenizer: FTS5Tokenizer {
-	/// 	var tokenizer: CFStringTokenizer!
-	/// 	var text: CFString!
-	///
-	/// 	required init(arguments: [String]) {
-	/// 		// Arguments not used
-	/// 	}
-	///
-	/// 	func set(text: String, reason: Database.FTS5TokenizationReason) {
-	/// 		// Reason not used
-	/// 		self.text = text as CFString
-	/// 		tokenizer = CFStringTokenizerCreate(kCFAllocatorDefault, self.text, CFRangeMake(0, CFStringGetLength(self.text)), kCFStringTokenizerUnitWord, nil)
-	/// 	}
-	///
-	/// 	func advance() -> Bool {
-	/// 		let nextToken = CFStringTokenizerAdvanceToNextToken(tokenizer)
-	/// 		guard nextToken != CFStringTokenizerTokenType(rawValue: 0) else {
-	/// 			return false
-	/// 		}
-	/// 		return true
-	/// 	}
-	///
-	/// 	func currentToken() -> String? {
-	/// 		let tokenRange = CFStringTokenizerGetCurrentTokenRange(tokenizer)
-	/// 		guard tokenRange.location != kCFNotFound /*|| tokenRange.length != 0*/ else {
-	/// 			return nil
-	/// 		}
-	/// 		return CFStringCreateWithSubstring(kCFAllocatorDefault, text, tokenRange) as String
-	/// 	}
-	///
-	/// 	func copyCurrentToken(to buffer: UnsafeMutablePointer<UInt8>, capacity: Int) throws -> Int {
-	/// 		let tokenRange = CFStringTokenizerGetCurrentTokenRange(tokenizer)
-	/// 		var bytesConverted = 0
-	/// 		let charsConverted = CFStringGetBytes(text, tokenRange, CFStringBuiltInEncodings.UTF8.rawValue, 0, false, buffer, capacity, &bytesConverted)
-	/// 		guard charsConverted > 0 else {
-	/// 			throw DatabaseError("Insufficient buffer size")
-	/// 		}
-	/// 		return bytesConverted
-	/// 	}
-	/// }
-	/// ```
-	///
-	/// - parameter name: The name of the tokenizer
-	/// - parameter type: The class implementing the tokenizer
-	///
-	/// - throws:  An error if the tokenizer can't be added
-	///
-	/// - seealso: [Custom Tokenizers](https://www.sqlite.org/fts5.html#custom_tokenizers)
-	public func addTokenizer<T: FTS5Tokenizer>(_ name: String, type: T.Type) throws {
-		// Fail early if FTS5 isn't available
-		let api_ptr = try get_fts5_api(for: db)
-
-		// Flesh out the struct containing the xCreate, xDelete, and xTokenize functions used by SQLite
-		var tokenizer_struct = fts5_tokenizer(xCreate: { (user_data, argv, argc, out) -> Int32 in
-			// Create the tokenizer instance using the creation function passed to fts5_api.xCreateTokenizer()
-			let args = UnsafeBufferPointer(start: argv, count: Int(argc))
-			let arguments = args.map { String(utf8String: $0.unsafelyUnwrapped).unsafelyUnwrapped }
-
-			let tokenizer: FTS5Tokenizer
-			do {
-				let creator = Unmanaged<FTS5TokenizerCreator>.fromOpaque(UnsafeRawPointer(user_data.unsafelyUnwrapped)).takeUnretainedValue()
-				tokenizer = try creator.construct(arguments)
-			}
-
-			catch let error {
-				os_log("Error constructing FTS5 tokenizer: %{public}@", type: .info, error.localizedDescription)
-				return SQLITE_ERROR
-			}
-
-			// tokenizer must live until the xDelete function is invoked; store it as a +1 object in ptr
-			let ptr = Unmanaged.passRetained(tokenizer as AnyObject).toOpaque()
-			out?.initialize(to: OpaquePointer(ptr))
-
-			return SQLITE_OK
-		}, xDelete: { p in
-			// Balance the +1 retain above
-			Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(p.unsafelyUnwrapped)).release()
-		}, xTokenize: { (tokenizer_ptr, context, flags, text_utf8, text_len, xToken) -> Int32 in
-			// Tokenize the text and invoke xToken for each token found
-			let tokenizer = Unmanaged<AnyObject>.fromOpaque(UnsafeRawPointer(tokenizer_ptr.unsafelyUnwrapped)).takeUnretainedValue() as! FTS5Tokenizer
-
-			// Set the text to be tokenized
-			let text = String(bytesNoCopy: UnsafeMutableRawPointer(mutating: text_utf8.unsafelyUnwrapped), length: Int(text_len), encoding: .utf8, freeWhenDone: false).unsafelyUnwrapped
-			let reason = FTS5TokenizationReason(flags)
-
-			tokenizer.setText(text, reason: reason)
-
-			// Use a local buffer for token extraction if possible
-			let bufsize = 512
-			var buf = UnsafeMutablePointer<UInt8>.allocate(capacity: bufsize)
-
-			defer {
-				buf.deallocate()
-			}
-
-			// Process each token and pass to FTS5
-			while tokenizer.advance() {
-				do {
-					// Attempt to copy the current token to buf
-					let byteCount = try tokenizer.copyCurrentToken(to: buf, capacity: bufsize)
-
-					let result = UnsafePointer(buf).withMemoryRebound(to: Int8.self, capacity: bufsize) { bytes in
-						return xToken.unsafelyUnwrapped(context, 0, bytes, Int32(byteCount), 0, Int32(byteCount))
-					}
-
-					guard result == SQLITE_OK else {
-						return result
-					}
-				}
-
-				catch {
-					// The token was too large to fit in buf
-					guard let token = tokenizer.currentToken() else {
-						continue
-					}
-					let utf8 = token.utf8
-					let result = xToken.unsafelyUnwrapped(context, 0, token, Int32(utf8.count), 0, Int32(utf8.count))
-					guard result == SQLITE_OK else {
-						return result
-					}
-				}
-			}
-
-			return SQLITE_OK
-		})
-
-		// user_data must live until the xDestroy function is invoked; store it as a +1 object
-		let user_data = FTS5TokenizerCreator { (args) -> FTS5Tokenizer in
-			return T(arguments: args)
-		}
-		let user_data_ptr = Unmanaged.passRetained(user_data).toOpaque()
-
-		guard api_ptr.pointee.xCreateTokenizer(UnsafeMutablePointer(mutating: api_ptr), name, user_data_ptr, &tokenizer_struct, { user_data in
-			// Balance the +1 retain above
-			Unmanaged<FTS5TokenizerCreator>.fromOpaque(UnsafeRawPointer(user_data.unsafelyUnwrapped)).release()
-		}) == SQLITE_OK else {
-			// xDestroy is not called if fts5_api.xCreateTokenizer() fails
-			Unmanaged<FTS5TokenizerCreator>.fromOpaque(user_data_ptr).release()
-			throw SQLiteError("Error creating FTS5 tokenizer", takingDescriptionFromDatabase: db)
-		}
-	}
-}
-
-extension Database.FTS5TokenizationReason {
-	/// Convenience initializer for conversion of `FTS5_TOKENIZE_` values
-	///
-	/// - parameter flags: The flags passed as the second argument of `fts5_tokenizer.xTokenize()`
-	init(_ flags: Int32) {
-		switch flags {
-		case FTS5_TOKENIZE_DOCUMENT: 						self = .document
-		case FTS5_TOKENIZE_QUERY: 							self = .query
-		case FTS5_TOKENIZE_QUERY | FTS5_TOKENIZE_PREFIX: 	self = .prefix
-		case FTS5_TOKENIZE_AUX: 							self = .aux
-		default:											preconditionFailure("Unexpected FTS5 flag")
-		}
 	}
 }
